requirements.txt
atproto
dotenv
python-dateutil
spacy
BeautifulSoup4
lxml
pillow